Add YAML policy file support (#77)
This change allows Aerleon to read policy files written in YAML as well as the traditional .pol format. It will search for files named \*.pol.yaml in the same places it would search for \*.pol files.

Include files can also be written in YAML. Cross-format includes are not allowed. A YAML file can include files named \*.pol-include.yaml while a .pol file can include files named \*.inc .

YAML files are loaded in PyYAML's "safe" mode.

Reviewers: please see [/doc/wiki/Yaml-files.md](https://github.com/ankben/aerleon/blob/yaml/doc/wiki/Yaml-files.md).

### How this works:

```
+----------+ 
| YAML file | --- load YAML ---> Native Python representation --- validation --> RawPolicy
+----------+                                                                        |
                                                                         BuiltinRecognizer.recognize()
+--------------+                                                                    v
| Policy Model |       <--- FromBuilder(PolicyBuilder) ---      RawPolicy (parsed, validated, normalized)
+--------------+ 
```

The ultimate goal of the YAML front-end is to produce a Policy model identical to what would be produced by an equivalent .pol file. Prior to this change the only way to construct (or even instantiate) a Policy model would be to run the .pol parser against a .pol file. This PR adds the PolicyBuilder class and policy.FromBuilder(builder) module function so that a Policy model can now be constructed directly from a native Python representation. The RawPolicy class acts as a flexible intermediate representation understood by PolicyBuilder.

The YAML front-end is responsible for loading the file into a RawPolicy - with minimal parsing and validation. In-depth parsing and validation of value syntax and semantics will be performed by the PolicyBuilder. Normalization of equivalent input types (e.g. string "100" vs integer 100) is also performed within PolicyBuidler.

Error messaging is a critical element of this change. Errors and warnings should be raised to the user on non-conforming YAML input. Those warnings should contain file and line context as much as possible. The current PR does have some limitations where validation errors generated within the PolicyBuilder lack file and line contexts. A follow-up PR to add spans and push them through to the builder phase would be desirable.

This change also adds a suite of standard value expression "recognizers" which match, parse and normalize common value expressions.