format all code with black/isort
> "Any customer can have a car painted any color that he wants so long as it
> is black".

Auto formatting source code is great for working in larger teams and allows
code review to focus on teh code and not formatting issues.

The defacto tool is "black" which is an opinonated formatting tool.  There
already seems to be configuration for black in the pyproject.toml file and
that wasn't modified.

https://black.readthedocs.io/en/stable/

In addition to "black" there is a tool called "isort" that sorts and formats
`import` lines again in the name of having a standard across all files.

https://pycqa.github.io/isort/

This PR is just running `black` and `isort` on the files and then adds a
formatting check to the github actions.