(re)format all code with black/isort
Seems like some files were either skipped when formatting black or have
drifted. This reformats all python files with black.

In addition to "black" there is a tool called "isort" that sorts and formats
`import` lines again in the name of having a standard across all files.

https://pycqa.github.io/isort/

This PR is just running `black` and `isort` on the files.  Another PR will  be
opened up update documentation and add a validation check on all PRs to
ensure they are properly formatted