add gh workflow to check for proper code formatting (+ docs)
This adds a check to make sure code is formated with black/isort on PR
or push to master (used existing main workflow and just added another job)

I renamed the workflow from main.yml to validate.yml to be more clear.

Also a minor update to the documentation to add a mention to isort and
how to configure your editor.