Use ProcessPoolExecutor for multiprocessing and fix logging
This switches to using a ProcessPoolExecutor which is a bit cleaner way
of handling a multiprocessing pool and also allows for RenderFile and
RenderACL to return it's results instead of appending to an existing
thread-safe list which is much more straight forward.

In addition logging was only being setup on the main process.  This
calls a setup_logging() function at the start of each multiprocessing
function to make sure all logging is done for all subprocesses (which
should make it match when multiprocessing isn't used).

This will allow a future (already written) PR to remove absl logging in
favor for the standar library `logging function`